Ok we got mdadm notifs!

+++ hosts/archiver/services.nix
@@ -196,7 +196,9 @@
               # if $dev does not exist, diskio_new_rw="", which will be safe
               diskio_new_rw="$(cat /sys/block/''${devices[$led]}/stat 2>/dev/null)"
               if [ "''${diskio_data_rw[$led]}" != "''${diskio_new_rw}" ]; then
-                  echo 1 > /sys/class/leds/$led/shot
+                  if [[ -f /sys/class/leds/$led/shot  ]]; then
+                      echo 1 > /sys/class/leds/$led/shot
+                  fi
               fi
               diskio_data_rw[$led]=$diskio_new_rw
           done
@@ -206,39 +208,40 @@
 
   };
 
-  services.opendkim.enable = true;
-  services.opendkim.domains = "csl:faceftw.dev";
-  services.opendkim.selector = "alerts._domainkey";
-  services.opendkim.user = "watchman";
-  services.opendkim.group = "watchman";
-  services.opendkim.settings.Mode = "sv";
-  services.opendkim.settings.Domain = "faceftw.dev";
-  services.opendkim.settings.RequireSafeKeys = "False";
-  services.opendkim.settings.ExternalIgnoreList = "127.0.0.1,::1,faceftw.dev";
-  services.opendkim.settings.InternalHosts = "127.0.0.1,::1";
+  # services.opendkim.enable = true;
+  # services.opendkim.domains = "csl:faceftw.dev";
+  # services.opendkim.selector = "alerts._domainkey";
+  # services.opendkim.user = "watchman";
+  # services.opendkim.group = "watchman";
+  # services.opendkim.settings.Mode = "sv";
+  # services.opendkim.settings.Domain = "faceftw.dev,archiver.localdomain";
+  # services.opendkim.settings.RequireSafeKeys = "False";
+  # services.opendkim.settings.ExternalIgnoreList = "127.0.0.1,::1,faceftw.dev";
+  # services.opendkim.settings.InternalHosts = "127.0.0.1,::1";
 
-  services.postfix.enable = true;
-  services.postfix.user = "watchman";
-  services.postfix.group = "watchman";
-  services.postfix.rootAlias = "archiver-alerts@faceftw.dev";
-  services.postfix.settings.main.relayhost = [ "faceftw.dev" ];
-  services.postfix.settings.main.smtp_tls_security_level = "dane";
-  services.postfix.settings.main.smtpd_milters = "unix:/run/opendkim/opendkim.sock";
-  services.postfix.settings.main.non_smtpd_milters = "$smtpd_milters";
-  services.postfix.settings.main.milter_default_action = "accept";
-  services.postfix.canonical = ''
-    root@archiver.localdomain archiver-alerts@faceftw.dev
-    @archiver.localdomain archiver-alerts@faceftw.dev
-    root@faceftw.dev archiver-alerts@faceftw.dev
-    @faceftw.dev archiver-alerts@faceftw.dev
-  '';
-  #   services.postfix.settings.main.smtp_generic_maps = "hash:/var/lib/postfix/conf/generic";
-  services.postfix.settings.main.myorigin = "faceftw.dev";
-  #   services.postfix.mapFiles.generic = ./postfix-generic;
+  # services.postfix.enable = true;
+  # services.postfix.user = "watchman";
+  # services.postfix.group = "watchman";
+  # services.postfix.rootAlias = "archiver-alerts@faceftw.dev";
+  # services.postfix.settings.main.relayhost = [ "faceftw.dev" ];
+  # services.postfix.settings.main.smtp_tls_security_level = "dane";
+  # services.postfix.settings.master.smtp.smtpd_milters = "unix:/run/opendkim/opendkim.sock";
+  # services.postfix.settings.master.non_smtpd_milters = "$smtpd_milters";
+  # services.postfix.settings.master.milter_default_action = "accept";
+  # services.postfix.canonical = ''
+  #   root@archiver.localdomain archiver-alerts@faceftw.dev
+  #   @archiver.localdomain archiver-alerts@faceftw.dev
+  #   root@faceftw.dev archiver-alerts@faceftw.dev
+  #   @faceftw.dev archiver-alerts@faceftw.dev
+  # '';
+  # services.postfix.extraAliases = "root: archiver-alerts@faceftw.dev";
+  # #   services.postfix.settings.main.smtp_generic_maps = "hash:/var/lib/postfix/conf/generic";
+  # services.postfix.settings.main.myorigin = "faceftw.dev";
+  # #   services.postfix.mapFiles.generic = ./postfix-generic;
 
-  #   environment.etc."postfix/generic".text = ''
-  #     root@archiver.local archiver-alerts@faceftw.dev
-  #     @archiver.local archiver-alerts@faceftw.dev
-  #   '';
+  # #   environment.etc."postfix/generic".text = ''
+  # #     root@archiver.local archiver-alerts@faceftw.dev
+  # #     @archiver.local archiver-alerts@faceftw.dev
+  # #   '';
 
 }
